Fixes for X86_64: - IPTR as storage variable for Get... functions - Variadic functions implemented with SLOWSTACK macros - sprintf implemented with VNewRawDoFmt rather than RawDoFmt - lots of ULONG to IPTR changes to get rid of warnings about   conversion to type with different size.
git-svn-id: https://svn.aros.org/svn/aros/trunk@33378 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/classes/datatypes/html/htmlclass.h
@@ -18,6 +18,6 @@ struct HtmlData
 	ULONG		style;
 	UBYTE		fgpen;
 	UBYTE		bgpen;
-	int		linenum;
+	long		linenum;
 };
